Fixing a snag in the library
that crashed https://github.com/MelodyLucid/freezing-ironman/commit/8870222fd6b6b0126c5fbe15e186737ac5062bd6

+++ components/runtime/src/miniboxing/runtime/MiniboxConversions.java
@@ -144,7 +144,7 @@ public class MiniboxConversions {
       case MiniboxConstants.FLOAT:
         return Float.floatToIntBits((java.lang.Float)a);
       case MiniboxConstants.DOUBLE:
-        return Double.doubleToLongBits((java.lang.Long)a);
+        return Double.doubleToLongBits((java.lang.Double)a);
       default:
         return (java.lang.Long)a;
     }
